refactor: extract feedback-panel.tsx into focused modules
Split the 1776-line feedback-panel.tsx into 5 focused files:
- feedback-utils.ts: types, constants, and useFeedbackData hook
- feedback-shared.tsx: shared UI helpers (FeedbackActions, RoundChip, etc.)
- feedback-mobile.tsx: MobileFeedbackSheet + MobileReviewSummarySheet
- review-summary-panel.tsx: ReviewSummaryPanel
- feedback-panel.tsx: ParentFeedbackPanel + FeedbackDetailPanel (665 lines)

Re-exports from feedback-panel.tsx maintain backward compatibility for
all existing consumers.
